tweak(header): Pass header content using children
This is also creates a provider to manage zap dialog state.
